Change (#118): move hardware-related features within Hanami
