+++ src/Hanami/Hanami.pro
@@ -225,7 +225,7 @@ HEADERS += \
     src/core/data_set_files/data_set_functions.h \
     src/core/data_set_files/image_data_set_file.h \
     src/core/data_set_files/table_data_set_file.h \
-    src/core/power_measuring.h \
+    src/hardware/power_measuring.h \
     src/core/processing/cluster_queue.h \
     src/core/processing/cluster_io_functions.h \
     src/core/processing/cpu/backpropagation.h \
@@ -236,9 +236,9 @@ HEADERS += \
     src/core/processing/processing_unit_handler.h \
     src/core/processing/section_update.h \
     src/core/routing_functions.h \
-    src/core/speed_measuring.h \
+    src/hardware/speed_measuring.h \
     src/core/temp_file_handler.h \
-    src/core/temperature_measuring.h \
+    src/hardware/temperature_measuring.h \
     src/core/thread_binder.h \
     src/core/value_container.h \
     src/database/audit_log_table.h \
@@ -335,13 +335,13 @@ SOURCES += \
     src/core/data_set_files/data_set_functions.cpp \
     src/core/data_set_files/image_data_set_file.cpp \
     src/core/data_set_files/table_data_set_file.cpp \
-    src/core/power_measuring.cpp \
+    src/hardware/power_measuring.cpp \
     src/core/processing/cluster_queue.cpp \
     src/core/processing/cpu_processing_unit.cpp \
     src/core/processing/processing_unit_handler.cpp \
-    src/core/speed_measuring.cpp \
+    src/hardware/speed_measuring.cpp \
     src/core/temp_file_handler.cpp \
-    src/core/temperature_measuring.cpp \
+    src/hardware/temperature_measuring.cpp \
     src/core/thread_binder.cpp \
     src/core/value_container.cpp \
     src/database/audit_log_table.cpp \
@@ -357,6 +357,3 @@ SOURCES += \
     src/database/users_table.cpp \
     src/hanami_root.cpp \
     src/main.cpp
-
-
-
