+++ src/Hanami/src/hardware/power_measuring.cpp
