+++ src/Hanami/src/hardware/temperature_measuring.cpp
