+++ src/Hanami/src/hardware/temperature_measuring.h
